Changed solr-query API.  It no longer supports result-type argument.
It turned out discarding the information by result-type argument
isn't feasible in practice, since the caller always need to check
if the result is pagenated or not.   Instead of returning a selected
information by result-type, we make solr-query always return the whole
SXML node, and provide separate functions to extract desired
information.